Don't mess around with XY()

+++ src/main/utils/code.js
@@ -57,7 +57,7 @@ int16_t gridIndexHorizontal(int16_t x, int16_t y) {
   return -1;
 }
 
-uint16_t XY(uint16_t x, uint16_t y) {
+uint16_t XY(uint8_t x, uint8_t y) {
   return gridIndexHorizontal(x, y);
 }
